Rename authorizer.ts -> lambdaAuthorizer.ts
